Kinvey setup now during application launch and accessed through application class
