+++ AndroidManifest.xml
@@ -27,6 +27,7 @@
         
     <application android:label="@string/app_name" 
         android:icon="@drawable/icon"
+        android:name=".ContactManagerApplication"
         android:allowBackup="true">
         <activity android:name=".ContactManager" android:label="@string/app_name">
             <intent-filter>
