+++ src/com/example/android/contactmanager/ContactAdderFragment.java
@@ -8,7 +8,6 @@ import android.accounts.AuthenticatorDescription;
 import android.accounts.OnAccountsUpdateListener;
 import android.content.ContentProviderOperation;
 import android.content.Context;
-import android.content.pm.PackageManager;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.provider.ContactsContract;
@@ -250,7 +249,7 @@ public class ContactAdderFragment extends Fragment implements OnAccountsUpdateLi
             String systemAccountType = a[i].type;
             AuthenticatorDescription ad = getAuthenticatorDescription(systemAccountType,
                     accountTypes);
-            AccountData data = new AccountData(a[i].name, ad);
+            AccountData data = new AccountData(getActivity(), a[i].name, ad);
             mAccounts.add(data);
         }
 
@@ -281,73 +280,6 @@ public class ContactAdderFragment extends Fragment implements OnAccountsUpdateLi
     private void updateAccountSelection() {
         // Read current account selection
         mSelectedAccount = (AccountData) mAccountSpinner.getSelectedItem();
-    }
-
-    /**
-     * A container class used to repreresent all known information about an account.
-     */
-    private class AccountData {
-        private String mName;
-        private String mType;
-        private CharSequence mTypeLabel;
-        private Drawable mIcon;
-
-        /**
-         * @param name The name of the account. This is usually the user's email address or
-         *        username.
-         * @param description The description for this account. This will be dictated by the
-         *        type of account returned, and can be obtained from the system AccountManager.
-         */
-        public AccountData(String name, AuthenticatorDescription description) {
-            mName = name;
-            if (description != null) {
-                mType = description.type;
-
-                // The type string is stored in a resource, so we need to convert it into something
-                // human readable.
-                String packageName = description.packageName;
-                PackageManager pm = getActivity().getPackageManager();
-
-                if (description.labelId != 0) {
-                    mTypeLabel = pm.getText(packageName, description.labelId, null);
-                    if (mTypeLabel == null) {
-                        throw new IllegalArgumentException("LabelID provided, but label not found");
-                    }
-                } else {
-                    mTypeLabel = "";
-                }
-
-                if (description.iconId != 0) {
-                    mIcon = pm.getDrawable(packageName, description.iconId, null);
-                    if (mIcon == null) {
-                        throw new IllegalArgumentException("IconID provided, but drawable not " +
-                                "found");
-                    }
-                } else {
-                    mIcon = getResources().getDrawable(android.R.drawable.sym_def_app_icon);
-                }
-            }
-        }
-
-        public String getName() {
-            return mName;
-        }
-
-        public String getType() {
-            return mType;
-        }
-
-        public CharSequence getTypeLabel() {
-            return mTypeLabel;
-        }
-
-        public Drawable getIcon() {
-            return mIcon;
-        }
-
-        public String toString() {
-            return mName;
-        }
     }
 
     /**
